refactor: componentized forms, soon will do same for fields too
