+++ frontend/src/routes/onboarding/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+    
+</script>
+
+hi
