+++ frontend/src/routes/settings/mediaserver/+page.svelte
@@ -1,27 +1,7 @@
 <script lang="ts">
 	import type { PageData } from './$types';
-	import { superForm } from 'sveltekit-superforms/client';
-	import { Button } from '$lib/components/ui/button';
-	import { Separator } from '$lib/components/ui/separator';
-	import { toast } from 'svelte-sonner';
-	import { Loader2 } from 'lucide-svelte';
-	import { page } from '$app/stores';
-	import clsx from 'clsx';
-	import * as Form from '$lib/components/ui/form';
-	import { mediaServerSettingsSchema, type MediaServerSettingsSchema } from '$lib/schemas/setting';
-	import { getContext } from 'svelte';
-
-	let formDebug: boolean = getContext('formDebug');
-
+	import MediaServerForm from '$lib/forms/media-server-form.svelte';
 	export let data: PageData;
-	const mediaServerForm = superForm(data.form);
-	const { form, message, delayed } = mediaServerForm;
-
-	$: if ($message && $page.status === 200) {
-		toast.success($message);
-	} else if ($message) {
-		toast.error($message);
-	}
 </script>
 
 <div class="flex flex-col">
@@ -30,52 +10,5 @@
 		Configure media server settings for Iceberg.
 	</p>
 
-	<Form.Root
-		schema={mediaServerSettingsSchema}
-		controlled
-		form={mediaServerForm}
-		let:config
-		debug={formDebug}
-	>
-		<div class="flex flex-col my-4 gap-4">
-			<Form.Field {config} name="plex_url">
-				<Form.Item class="flex flex-col md:flex-row items-start md:items-center max-w-6xl">
-					<Form.Label
-						class="text-base md:text-lg font-semibold w-48 min-w-48 text-muted-foreground"
-					>
-						Plex URL
-					</Form.Label>
-					<Form.Input class="text-sm md:text-base" spellcheck="false" />
-				</Form.Item>
-				<Form.Validation class="text-sm md:text-base text-red-500" />
-			</Form.Field>
-
-			<Form.Field {config} name="plex_token">
-				<Form.Item class="flex flex-col md:flex-row items-start md:items-center max-w-6xl">
-					<Form.Label
-						class="text-base md:text-lg font-semibold w-48 min-w-48 text-muted-foreground"
-					>
-						Plex Token
-					</Form.Label>
-					<Form.Input
-						class={clsx('transition-all duration-300 text-sm md:text-base', {
-							'blur-sm hover:blur-none focus:blur-none': $form.plex_token.length > 0
-						})}
-						spellcheck="false"
-					/>
-				</Form.Item>
-				<Form.Validation class="text-sm md:text-base text-red-500" />
-			</Form.Field>
-
-			<Separator class=" mt-4" />
-			<div class="flex w-full justify-end">
-				<Button disabled={$delayed} type="submit" size="sm" class="w-full md:max-w-max">
-					{#if $delayed}
-						<Loader2 class="w-4 h-4 animate-spin mr-2" />
-					{/if}
-					Save changes
-				</Button>
-			</div>
-		</div>
-	</Form.Root>
+	<MediaServerForm data={data.form} />
 </div>
